+++ src/raytracer/PointLight.cpp
@@ -1,6 +1,6 @@
 #include "PointLight.h"
 
-PointLight::PointLight(const glm::vec4 &position, float intensity) {
+PointLight::PointLight(const glm::dvec4 &position, double intensity) {
     m_position = position;
     m_intensity = intensity;
 }
