+++ src/raytracer/Ray.cpp
@@ -1,15 +1,15 @@
 #include "Ray.h"
 
-Ray::Ray(glm::vec3 point, glm::vec3 direction) {
-    m_origin = glm::vec4(point, 1.0);
-    m_direction = glm::vec4(direction, 0.0);
+Ray::Ray(glm::dvec3 point, glm::dvec3 direction) {
+    m_origin = glm::dvec4(point, 1.0);
+    m_direction = glm::dvec4(direction, 0.0);
 }
 
-glm::vec4 Ray::at(float distance) const {
+glm::dvec4 Ray::at(double distance) const {
     return m_origin + m_direction * distance;
 }
 
-Ray Ray::transform(glm::mat4 transformation) const {
+Ray Ray::transform(glm::dmat4 transformation) const {
     auto position = transformation * m_origin;
     auto direction = transformation * m_direction;
     return { position, direction };
