+++ src/raytracer/Sphere.h
@@ -13,20 +13,20 @@ public:
     Sphere();
 
     Intersections intersect(const Ray &ray) const;
-    Sphere &withTransformation(const glm::mat4 &transformation);
+    Sphere &withTransformation(const glm::dmat4 &transformation);
     Sphere &withMaterial(const Material &material);
-    glm::mat4 getTransformation();
+    glm::dmat4 getTransformation();
 
-    glm::vec4 getNormalAt(const glm::vec4 &point) const;
+    glm::dvec4 getNormalAt(const glm::dvec4 &point) const;
     Color getColor() const { return m_color; }
     Material getMaterial() const { return m_material; }
 
 private:
     Material m_material;
-    glm::vec4 m_position { 0, 0, 0, 1 };
-    glm::mat4 m_model { glm::identity<glm::mat4>() };
-    glm::mat4 m_modelInverse { glm::identity<glm::mat4>() };
-    glm::mat4 m_modelInverseTranspose { glm::identity<glm::mat4>() };
+    glm::dvec4 m_position { 0, 0, 0, 1 };
+    glm::dmat4 m_model { glm::identity<glm::dmat4>() };
+    glm::dmat4 m_modelInverse { glm::identity<glm::dmat4>() };
+    glm::dmat4 m_modelInverseTranspose { glm::identity<glm::dmat4>() };
     Color m_color { 1, 1, 1, 1 };
 };
